기타 스타일 수정
