+++ src/SubPage/SubPageStyle.js
@@ -1,14 +1,14 @@
 import styled from 'styled-components';
-import {Link as LinkR} from 'react-router-dom';
+// import {Link as LinkR} from 'react-router-dom';
 import Image from './overlay-bg.jpg';
-import Image2 from './Building.jpg';
-import Image3 from './jellyFish.jpg';
+import Image2 from './Image/Building.jpg';
+import Image3 from './Image/jellyFish.jpg';
 
 export const SubPageWrapper = styled.div`
   width:100%;
   height:100vh;
   background:blue;
-  z-index:999;
+  z-index:-999;
 //   opacity:100%;
  overflow:hidden;
 
@@ -24,125 +24,23 @@ border:1px solid #000;
 background:blue;
 opacity:80%;
 position:relative;
-z-index:1;
+z-index:999;
  
 `;
-export const SubPageHeader = styled.div`
-   width:100%;
-   height:44px;
-   padding-top:20px;
-   padding-bottom:20px;
-   border:1px solid #000;
-   // background:url(${Image}) no-repeat;
-   background:url(${Image}) no-repeat; 
-   background-size:cover;
-   
-   color:#000;
-   z-index:1;
-   `;
-   
-   export const SubPageContainer = styled.div`
-   max-width:1296px;
-   hieght:44px;
-   margin:auto;
-   //    border:1px solid red;
-   padding-left:12px;
-   display:flex;
-   padding:right:12px;
-   // background:red;
-   z-index:999;
  
-  
-`;
 
-export const SubPageLogo = styled.div`
-  box-sizing:border-box;
-  color:#fff;
-  z-index:999;
-
-`
-
-export const SubPageLogoLink = styled(LinkR)`
-  text-decoration:none;
-  color:#fff;
-  font:bold 18px/24px arial;
-  display:block;
-  margin-left:20px;
-  float:left;
-  z-index:999;
-  color:green;
- 
-  `;
-  
-  export const SubPageNavbar = styled.div`
-  max-width:691.750px;
-  height:44px;
-//   border:1px solid #000;
-  color:#fff;
-  float:right;
-  position:relative;left:50%;
-  bottom:5px;
-  z-index:999;
-:;
-   
-  `;
-  
-  export const SubPageNavUl = styled.div`
-  max-width:691.750px;
-  list-stlye:none;
-  display:flex;
-  z-index:999;
-  
-  `
-  
-  export const SubPageNavList = styled(LinkR)`
-  float:right;
-   margin-left:20px;
-   padding:10px 20px 10px 0px;
-  color:#000;
-   text-decoration:none;
-   z-index:999;
+export const SubVisualContainer = styled.div`
+width:1320px;
+height:563.77px;
+border:1px solid #000;
+margin:auto;
 `;
 
 
 
 
-export const SubPageBackHeader = styled.div`
-//  :;
-//    background:#000;  
-   box-sizing:border-box;
-   max-width:100%;
-   height:102px;
-   border:1px solid #000;
-   position:relative;
-   top:-150%;
-   //  background:red;
-   `;
-   
-   export const SubPageBackTable = styled.div`
-   max-width:100%;
-   height:102px;
-   border:1px solid #000;
-   color:#000;
-   // background:red;
-   `;
-   
-   export const SubPageBackContainer = styled.div`
-   max-width:1320px;
-  height:86px;
-  margin:auto;
-`;
 
-export const SubIntroTitle = styled.div`
- 
-`
 
-export const SubPageBackContainerH1 = styled.div`
-   color:#000;
-   position:relative;
-   left:40%;
-   top:20%;
-`;
 
 export const SubVisual = styled.div`
   width:100%;
@@ -158,12 +56,6 @@ export const SubVisualSection = styled.div`
 
 `;
 
-export const SubVisualContainer = styled.div`
-   width:1320px;
-   height:563.77px;
-   border:1px solid #000;
-   margin:auto;
-`;
 
 export const SubVisualRow = styled.div`
    max-width:1320px;
@@ -187,7 +79,7 @@ height:539.77px;
 background:url(${Image}) no-repeat;
 background-size:cover;
 // border:1px solid blue;
-z-index:1;
+z-index:999;
 `
 
 export const BackgroundBg2 = styled.div`
@@ -310,4 +202,7 @@ export const SubVisualCreadits = styled.div`
   margin:auto;
   font:bold 14px/24px arial;
   color:#fff;
+` 
+export const SubIntroTitle = styled.div`
+ 
 `
