+++ src/SubPage/style/SlideShow.scss
@@ -2,11 +2,12 @@
     border:1px solid #000;
     max-width:856px;
     height:539.77px;
-
+    z-index:1;
 }
 
 .SliderForm {
     border: 1px solid red;
     max-width:856px;
     height:539.77px;
+    z-index:1;
 }
